feat(types): changes "condition" property to "conditions" (array)
BREAKING CHANGE: The "condition" type has been changed to be an array

This means that from now on, you'll need to provide a single or multiple conditions in an array.
The conditions logic is based on an "OR" condition, so only one of the conditions needs to be true.
If all of the conditions are false, then the flag is returning false.